v5/requirements.txt
numpy>=1.23
PyYAML>=6.0
opencv-python>=4.8
ultralytics>=8.0
